build: disable FP units
Even though MMX and SSE are perfectly fine to be used on UEFI, there
exist many firmwares in the wild that incorrectly initialize (or don't
initialize at all) the FP co-processors, causing exceptions to be
thrown at random memory addresses used as exception handler.

Lets be on the safe side and just disable both. You can easily override
it in your project, if you really want it.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>